farmer: Move rpc_client_* files to a new folder.
